refactor: mode + command handling
