+++ server/env.ts
@@ -14,28 +14,6 @@ export const env = createEnv({
     AUTO_ADD_CHANNEL: z.string().optional(),
     // Channel required for keywords / relevance check
     OPT_IN_CHANNEL: z.string().optional(),
-    // Comma-separated or JSON array of channel IDs gork will never respond in
-    BLOCKED_CHANNELS: z.preprocess((val) => {
-      if (typeof val === 'string') {
-        const s = val.trim();
-        if (s === '') {
-          return;
-        }
-        try {
-          const parsed = JSON.parse(s);
-          if (Array.isArray(parsed)) {
-            return parsed;
-          }
-        } catch {
-          // ignore
-        }
-        return s
-          .split(',')
-          .map((p) => p.trim())
-          .filter(Boolean);
-      }
-      return val;
-    }, z.array(z.string()).optional()),
     // Channel for report notifications
     REPORTS_CHANNEL: z.string().optional(),
     // Channel for logs: strikes, bans, bot join/leave (e.g. #gork-logs)
